chore: move retry logic out of sql query to its own file
