+++ apps/passport-server/src/util/retry.ts
@@ -1,0 +1,33 @@
+import { getActiveSpan } from "@opentelemetry/api/build/src/trace/context-utils";
+
+export async function execWithRetry<T>(
+  task: () => Promise<T>,
+  shouldRetry: (e: unknown) => boolean,
+  maxTries: number
+): Promise<T> {
+  const span = getActiveSpan();
+  const backoffFactorMs = 100;
+  let latestError = undefined;
+
+  span?.setAttribute("max_tries", maxTries);
+
+  for (let i = 0; i < maxTries; i++) {
+    span?.setAttribute("try_count", i + 1);
+
+    try {
+      const result = await task();
+      span?.setAttribute("retry_succeeded", true);
+      return result;
+    } catch (e) {
+      latestError = e;
+      if (!shouldRetry(e)) {
+        break;
+      }
+    }
+
+    await sleep(backoffFactorMs * (i + 1));
+  }
+
+  span?.setAttribute("retry_succeeded", false);
+  throw latestError;
+}
